Stabalized Tree with Springs and Added Wind
Added some Strings between the Joints of the Tree and the Position where they originally were created. This pulls the Joints always back to the original position and so stabalizes the Structure of the tree. The Strength of the Springs decreases further out.

Also add randomly occuring wind that pushes the tree from left to right